app/api/requirements.txt
fastapi==0.100.0 
uvicorn==0.23.0
sqlalchemy==2.0.19
python-dotenv==1.0.0
psycopg2==2.9.6